Make BPM usage configurable
The usage of BPM in cube-release was made configurable by reintroducing
the control scripts for registry and sync as well as a new control script
for stager. An additional operations file was added in case you do not
want to use BPM, for example in combination with fissile.

Also, we fixed the ERB for --skipSslValidation in stager in sync control
scripts. You cannot use `--skipSslValidation true`, but only the argument
itself or omit it if not needed.

Tested against BOSH Lite.

Signed-off-by: James Tuddenham <TUDDENHA@uk.ibm.com>